+++ examples/keras/simple_mnist_epochtrainer.py
@@ -81,9 +81,12 @@ trainer = MNISTTrainer()
 
 tuning = letstune.tune(
     trainer,
-    16,
+    8,
     results_dir=Path.home() / "ltexamples/keras/simple_mnist_epochtrainer",
-    training_maximum_duration=timedelta(seconds=30),
+    rounds=[
+        timedelta(seconds=2),
+        timedelta(seconds=4),
+    ],
 )
 print(f" DONE: {tuning}")