+++ examples/keras/simple_mnist_kerastrainer.py
@@ -69,10 +69,13 @@ trainer = letstune.keras.KerasTrainer(
 
 tuning = letstune.tune(
     trainer,
-    16,
+    8,
     dataset=(x_train, y_train, (x_val, y_val)),
     results_dir=Path.home() / "ltexamples/keras/simple_mnist_kerastrainer",
-    training_maximum_duration=timedelta(seconds=30),
+    rounds=[
+        timedelta(seconds=2),
+        timedelta(seconds=4),
+    ],
 )
 print(f" DONE: {tuning}")